modularize the routes
